feat: improved onboarding
